% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{refresh_proxy_list}
\alias{refresh_proxy_list}
\title{Refresh the Proxy List.}
\usage{
refresh_proxy_list(force = FALSE)
}
\arguments{
\item{force}{(Default FALSE); This function will do nothing if the proxy
table has been refreshed in the past hour, unless `force = TRUE`.}
}
\description{
Fetch a new proxy table via the API provided by Geonode (see
  https://geonode.com/free-proxy-list/) and update the proxy list used for
  scraping.
}
