% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{user_agents}
\alias{user_agents}
\title{User Agents.}
\format{
## `user_agents`
  A character vector of common user agents.
}
\source{
<https://developers.whatismybrowser.com/useragents/explore/hardware_type_specific/computer/>
}
\usage{
user_agents
}
\description{
A list of user agents to cycle when scraping.
}
\keyword{datasets}
